Now initcalls return error codes and have names
What next, giving initcalls voting rights?

+++ include/sprintf.h
@@ -6,9 +6,9 @@
 #include <stdarg.h>
 #include <stddef.h>
 
-int snprintf(char *restrict s, size_t n, const char *restrict fmt, ...) __format(printf, 3, 4) __weak;
-int sprintf(char *restrict s, const char *restrict fmt, ...) __format(printf, 2, 3) __weak;
-int vsnprintf(char *restrict s, size_t n, const char *restrict fmt, va_list ap) __weak;
-int vsprintf(char *restrict s, const char *restrict fmt, va_list ap) __weak;
+size_t snprintf(char *restrict s, size_t n, const char *restrict fmt, ...) __format(printf, 3, 4) __weak;
+size_t sprintf(char *restrict s, const char *restrict fmt, ...) __format(printf, 2, 3) __weak;
+size_t vsnprintf(char *restrict s, size_t n, const char *restrict fmt, va_list ap) __weak;
+size_t vsprintf(char *restrict s, const char *restrict fmt, va_list ap) __weak;
 
 #endif /* __INCLUDE_PSPRINTF_H__ */
