+++ kernel/lib/snprintf.c
@@ -2,9 +2,9 @@
 /* Copyright (c), 2022, Kaneru Contributors */
 #include <sprintf.h>
 
-int snprintf(char *restrict s, size_t n, const char *restrict fmt, ...)
+size_t snprintf(char *restrict s, size_t n, const char *restrict fmt, ...)
 {
-    int r;
+    size_t r;
     va_list ap;
     va_start(ap, fmt);
     r = vsnprintf(s, n, fmt, ap);
