+++ kernel/lib/vsprintf.c
@@ -3,7 +3,7 @@
 #include <sprintf.h>
 #include <stdint.h>
 
-int vsprintf(char *restrict s, const char *restrict fmt, va_list ap)
+size_t vsprintf(char *restrict s, const char *restrict fmt, va_list ap)
 {
     return vsnprintf(s, SIZE_MAX, fmt, ap);
 }
